Changed the plural controller names into singular
