+++ app/Http/Controllers/RecipeTagController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Ingredients;
+use App\Models\RecipeTags;
 use Illuminate\Http\Request;
 
-class IngredientsController extends Controller
+class RecipeTagController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -34,7 +34,7 @@ class IngredientsController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Ingredients $ingredients)
+    public function show(RecipeTags $recipeTags)
     {
         //
     }
@@ -42,7 +42,7 @@ class IngredientsController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Ingredients $ingredients)
+    public function edit(RecipeTags $recipeTags)
     {
         //
     }
@@ -50,7 +50,7 @@ class IngredientsController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Ingredients $ingredients)
+    public function update(Request $request, RecipeTags $recipeTags)
     {
         //
     }
@@ -58,7 +58,7 @@ class IngredientsController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Ingredients $ingredients)
+    public function destroy(RecipeTags $recipeTags)
     {
         //
     }
