+++ app/Http/Controllers/TagController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Users;
+use App\Models\Tags;
 use Illuminate\Http\Request;
 
-class UsersController extends Controller
+class TagController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -34,7 +34,7 @@ class UsersController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Users $users)
+    public function show(Tags $tags)
     {
         //
     }
@@ -42,7 +42,7 @@ class UsersController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Users $users)
+    public function edit(Tags $tags)
     {
         //
     }
@@ -50,7 +50,7 @@ class UsersController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Users $users)
+    public function update(Request $request, Tags $tags)
     {
         //
     }
@@ -58,7 +58,7 @@ class UsersController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Users $users)
+    public function destroy(Tags $tags)
     {
         //
     }
